tree_drive/requirements.txt
astroid==2.3.2
autopep8==1.4.4
cffi==1.13.2
cryptography==2.8
Django==2.2.7
django-rest-knox==4.1.0
djangorestframework==3.10.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pkg-resources==0.0.0
pycodestyle==2.5.0
pycparser==2.19
pylint==2.4.3
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
